fix slots data type

+++ server/src/main/java/org/apache/iotdb/db/mpp/common/header/HeaderConstant.java
@@ -218,7 +218,7 @@ public class HeaderConstant {
                 new ColumnHeader(COLUMN_TYPE, TSDataType.TEXT),
                 new ColumnHeader(COLUMN_STATUS, TSDataType.TEXT),
                 new ColumnHeader(COLUMN_STORAGE_GROUP, TSDataType.TEXT),
-                new ColumnHeader(COLUMN_SLOTS, TSDataType.INT32),
+                new ColumnHeader(COLUMN_SLOTS, TSDataType.INT64),
                 new ColumnHeader(COLUMN_DATANODE_ID, TSDataType.INT32),
                 new ColumnHeader(COLUMN_HOST_ADDRESS, TSDataType.TEXT),
                 new ColumnHeader(COLUMN_PORT, TSDataType.INT32)),
